修改bug

+++ src/main/java/com/hgys/iptv/service/impl/AccountSettlementServiceImpl.java
@@ -547,7 +547,7 @@ public class AccountSettlementServiceImpl implements AccountSettlementService {
             }else if (3 == vm.getSet_type()){
                 //3、先删除再新增产品级结算结算源数据
                 //单维度
-                if (!vm.getDimensionAddVM().isEmpty()){
+                if (null != vm.getDimensionAddVM() && !vm.getDimensionAddVM().isEmpty()){
                     settlementProductSingleRepository.deleteByMasterCode(code);
 
                     List<OrderProductDimensionAddVM> dimensionAddVM = vm.getDimensionAddVM();
@@ -558,7 +558,7 @@ public class AccountSettlementServiceImpl implements AccountSettlementService {
                         single.setMasterCode(code);
                         settlementProductSingleRepository.save(single);
                     }
-                }else if (!vm.getDimensionListAddVMS().isEmpty()){
+                }else if (null != vm.getDimensionListAddVMS() && !vm.getDimensionListAddVMS().isEmpty()){
                     //先删除之前数据
                     settlementProductManyRepository.deleteByMasterCode(code);
